Support React Native by adding check for window.document
React Native polyfills the `window` global, but it's not an actual DOM
`window` object. Thus, checking `typeof window !== 'undefined'` is not
enough for asserting that it's an *actual* browser environment. Add a
`window.document != null` check and abstract out into an `isBrowser()`
function.

I created the minified file using UglifyJS2.